add workspace CRUD

+++ internal/models/workspace.go
@@ -3,9 +3,9 @@ package models
 import "time"
 
 type Workspace struct {
-	ID        int       `json:"id"`
+	ID        int64     `json:"id"`
 	Name      string    `json:"name"`
-	OwnerID   int       `json:"owner_id"`
+	OwnerID   int64     `json:"owner_id"`
 	CreatedAt time.Time `json:"created_at"`
 	UpdatedAt time.Time `json:"updated_at"`
 }
